Move actions to a restful setup in TenancyController
