+++ LBHTenancyAPI/Controllers/TenanciesController.cs
@@ -13,10 +13,14 @@ namespace LBHTenancyAPI.Controllers
     public class TenanciesController : Controller
     {
         private readonly IListTenancies listTenancies;
+        private readonly IListAllPayments listAllPayments;
+        private readonly IListAllArrearsActions listAllArrearsActions;
 
-        public TenanciesController(IListTenancies listTenancies)
+        public TenanciesController(IListTenancies listTenancies, IListAllArrearsActions listAllArrearsActions, IListAllPayments listAllPayments)
         {
             this.listTenancies = listTenancies;
+            this.listAllArrearsActions = listAllArrearsActions;
+            this.listAllPayments = listAllPayments;
         }
 
         [HttpGet]
@@ -44,6 +48,53 @@ namespace LBHTenancyAPI.Controllers
             var result = new Dictionary<string, object>
             {
                 {"tenancies", tenancies}
+            };
+
+            return Ok(result);
+        }
+
+
+
+        [HttpGet]
+        [Route("{tenancyRef}/payments")]
+        public async Task<IActionResult> PaymentTransactionDetails(string tenancyRef)
+        {
+            var response = listAllPayments.Execute(tenancyRef);
+            var paymentsTransaction = response.PaymentTransactions.ConvertAll(paymentTrans => new Dictionary<string, object>
+            {
+                {"ref", paymentTrans.Ref},
+                {"amount", paymentTrans.Amount},
+                {"date", paymentTrans.Date},
+                {"type", paymentTrans.Type},
+                {"property_ref", paymentTrans.PropertyRef}
+            });
+
+            var result = new Dictionary<string, object>
+            {
+                {"payment_transactions", paymentsTransaction}
+            };
+
+            return Ok(result);
+        }
+
+        [HttpGet]
+        [Route("{tenancyRef}/actions")]
+        public async Task<IActionResult> GetActionDiaryDetails(string tenancyRef)
+        {
+            var response = listAllArrearsActions.Execute(tenancyRef);
+            var arrearActionDiary = response.ActionDiaryEntries.ConvertAll(actionDiary => new Dictionary<string, object>
+            {
+                {"balance", actionDiary.Balance},
+                {"code", actionDiary.Code},
+                {"code_name", actionDiary.CodeName},
+                {"date", actionDiary.Date.ToString()},
+                {"comment", actionDiary.Comment},
+                {"universal_housing_username", actionDiary.UniversalHousingUsername}
+            });
+
+            var result = new Dictionary<string, object>
+            {
+                {"arrears_action_diary_events", arrearActionDiary}
             };
 
             return Ok(result);
