+++ LBHTenancyAPITest/Test/UseCases/ListAllPaymentsForTenancyTest.cs
@@ -51,7 +51,7 @@ namespace LBHTenancyAPITest.Test.UseCases
                 {
                     new AllPaymentsForTenancy.PaymentTransaction
                     {
-                        Ref= payment.TransactionsRef,
+                        Ref= payment.TransactionRef,
                         Amount= payment.TransactionAmount.ToString("C"),
                         Date = string.Format("{0:u}", payment.TransactionDate),
                         Type = payment.TransactionType,
